third_party/blink/web_tests/wpt_internal/html/canvas/drawElement/allow-readback/html-images-ignored.https.sub.html

<!DOCTYPE HTML>
<html>
<head>
  <title>drawElement with allowReadback does not draw cross-origin HTML images</title>
  <script src='/resources/testharness.js'></script>
  <script src='/resources/testharnessreport.js'></script>
  <style>
  #child {
    width: 100px;
    height: 200px;
    background: green;
  }
  #sameOrigin {
    position: absolute;
    left: 0px;
    top: 0px;
  }
  #crossOrigin {
    position: absolute;
    left: 0px;
    top: 100px;
  }
  </style>
</head>
<body>
  <canvas id=canvas width="100" height="200" layoutsubtree>
    <div id=child>
      <img id=sameOrigin width=100 height=100 src="https://{{location[host]}}/css/css-masking/mask-image/support/red-100x100.png"/>
      <img id=crossOrigin width=100 height=100 src="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/css/css-masking/mask-image/support/red-100x100.png"/>
    </div>
  </canvas>

  <script>
    let sPendingImagesToLoad = 2;

    function pendingImageLoaded() {
      if (!--sPendingImagesToLoad) {
        test(function(t) {
          var context = canvas.getContext("2d");
          context.drawElement(child, 0, 0, { allowReadback: true });

          pixel = context.getImageData(50, 50, 1, 1).data;
          assert_equals(pixel[0], 255, "Same origin should draw");
          assert_equals(pixel[1], 0, "Same origin should draw");
          assert_equals(pixel[2], 0, "Same origin should draw");

          pixel = context.getImageData(50, 150, 1, 1).data;
          assert_equals(pixel[0], 0, "Cross origin should not draw");
          assert_equals(pixel[1], 128, "Cross origin should not draw");
          assert_equals(pixel[2], 0, "Cross origin should not draw");

          t.done();
        });
      }
    }

    sameOrigin.onload = pendingImageLoaded;
    crossOrigin.onload = pendingImageLoaded;
  </script>
</body>
</html>
